requirements.txt
matplotlib==3.3.3
numpy==1.17.0
reportlab==3.5.67
mlflow==1.14.1
torchsummary==1.5.1
scipy==1.5.4
tqdm==4.32.1
numba==0.45.1
tensorflow_model_optimization==0.4.0
torchpack==0.3.0
svglib==1.1.0
torch==1.8.1+cu111
torchvision==0.9.1+cu111
Pillow==8.2.0
PyYAML==5.4.1
scikit_learn==0.24.2
tensorboardX==2.2
tensorflow==2.4.1
